# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      requestId:
        type: string
    type: object
...
  dto.ListNotificationsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.NotificationLogResponse'
        type: array
      nextCursor:
        type: string
    type: object
...
  dto.NotificationLogResponse:
    properties:
      body:
        type: string
      createdAt:
        type: string
      id:
        type: string
      leadMins:
        type: integer
      readAt:
        type: string
      referenceId:
        type: string
      scheduledFor:
        type: string
      sentAt:
        type: string
      status:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  dto.NotificationPreferencesResponse:
    properties:
      dailyDigestEnabled:
        type: boolean
      dailyDigestHour:
        type: integer
      eventAtTime:
        type: boolean
      eventLeadMins:
        items:
          type: integer
        type: array
      eventsEnabled:
        type: boolean
      quietEnd:
        description: '"HH:MM"'
        type: string
      quietHoursEnabled:
        type: boolean
      quietStart:
        description: '"HH:MM"'
        type: string
      reminderAtTime:
        type: boolean
      reminderLeadMins:
        items:
          type: integer
        type: array
      remindersEnabled:
        type: boolean
      routineAtTime:
        type: boolean
      routineLeadMins:
        items:
          type: integer
        type: array
      routinesEnabled:
        type: boolean
      taskAtTime:
        type: boolean
      taskLeadMins:
        items:
          type: integer
        type: array
      tasksEnabled:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  dto.UpdateNotificationPreferencesRequest:
    properties:
      dailyDigestEnabled:
        type: boolean
      dailyDigestHour:
        type: integer
      eventAtTime:
        type: boolean
      eventLeadMins:
        items:
          type: integer
        type: array
      eventsEnabled:
        type: boolean
      quietEnd:
        type: string
      quietHoursEnabled:
        type: boolean
      quietStart:
        type: string
      reminderAtTime:
        type: boolean
      reminderLeadMins:
        items:
          type: integer
        type: array
      remindersEnabled:
        type: boolean
      routineAtTime:
        type: boolean
      routineLeadMins:
        items:
          type: integer
        type: array
      routinesEnabled:
        type: boolean
      taskAtTime:
        type: boolean
      taskLeadMins:
        items:
          type: integer
        type: array
      tasksEnabled:
        type: boolean
    type: object
...
info:
  contact: {}
  description: API do MVP Organiq.
  title: Organiq API
  version: "0.1"
paths:
...
  /v1/notification-preferences:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NotificationPreferencesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar preferências de notificação
      tags:
      - Notifications
    put:
      consumes:
      - application/json
      parameters:
      - description: Update preferences request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateNotificationPreferencesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NotificationPreferencesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar preferências de notificação
      tags:
      - Notifications
  /v1/notifications:
    get:
      parameters:
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListNotificationsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar notificações
      tags:
      - Notifications
  /v1/notifications/{id}/read:
    patch:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Marcar notificação como lida
      tags:
      - Notifications
  /v1/notifications/read-all:
    patch:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Marcar todas como lidas
      tags:
      - Notifications
  /v1/notifications/test:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Enviar notificação de teste
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    description: 'Formato: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"